chore: rename user endpoint to auth endpoint since it better represents what its doing
